Kickstarter campaign app

+++ blockchain/ethereum-solidity/campaign-ido/ethereum/contracts/Campaign.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.4.17;
 contract CampaignFactory {
     address[] public deployedCampaigns;
 
-    function createCampaign(uint256 minimum) public {
+    function createCampaign(uint minimum) public {
         address newCampaign = new Campaign(minimum, msg.sender);
         deployedCampaigns.push(newCampaign);
     }
@@ -16,25 +16,25 @@ contract CampaignFactory {
 contract Campaign {
     struct Request {
         string description;
-        uint256 value;
+        uint value;
         address recipient;
         bool complete;
-        uint256 approvalCount; // This is stored separately because mapping does not store keys and we wouldn't be able to do e.g Object.keys(mapping).length
+        uint approvalCount; // This is stored separately because mapping does not store keys and we wouldn't be able to do e.g Object.keys(mapping).length
         mapping(address => bool) approvals;
     }
 
     Request[] public requests;
     address public manager;
-    uint256 public minimumContribution;
+    uint public minimumContribution;
     mapping(address => bool) public approvers;
-    uint256 public approversCount;
+    uint public approversCount;
 
     modifier restrictToManager {
         require(msg.sender == manager);
         _;
     }
 
-    function Campaign(uint256 minimum, address creator) public {
+    function Campaign(uint minimum, address creator) public {
         manager = creator;
         minimumContribution = minimum;
     }
@@ -48,7 +48,7 @@ contract Campaign {
 
     function createRequest(
         string description,
-        uint256 value,
+        uint value,
         address recipient
     ) public restrictToManager {
         Request memory newRequest =
@@ -64,7 +64,7 @@ contract Campaign {
         requests.push(newRequest);
     }
 
-    function approveRequest(uint256 index) public {
+    function approveRequest(uint index) public {
         Request storage request = requests[index]; // we use storage because we actually want to mutate the data in the storage on the network.
 
         require(approvers[msg.sender]);
@@ -74,7 +74,7 @@ contract Campaign {
         request.approvalCount++;
     }
 
-    function finalizeRequest(uint256 index) public restrictToManager {
+    function finalizeRequest(uint index) public restrictToManager {
         Request storage request = requests[index];
         require(!request.complete);
         require(request.approvalCount > (approversCount / 2));
@@ -87,10 +87,10 @@ contract Campaign {
         public
         view
         returns (
-            uint256,
-            uint256,
-            uint256,
-            uint256,
+            uint,
+            uint,
+            uint,
+            uint,
             address
         )
     {
@@ -103,7 +103,7 @@ contract Campaign {
         );
     }
 
-    function getRequestsCount() public view returns (uint256) {
+    function getRequestsCount() public view returns (uint) {
         return requests.length;
     }
 }
